Lambda Custom Integration (Not Proxy)
This commit shows a working version of an API that uses non-proxy λ
integration.  It uses VTL templates to parse standard raised exceptions.

It may be usefull if you want to decouple the code from the API and thus
raise Exceptions without having to have a 'catch-all' at the top level
to ensure that the correct output object is sent to a lambda-proxy APIG
resource.

This also allows message headers and other request or response
modification outside the λ and may improve efficiency/reduce cost by
rejecting bad responses before they hit the λ although this is not
demonstrated in this version.

Some would argue that dealing with the explicit response including
headers and statusCode in the λ code allows for the interface to be
tested by local unit tests.  The same argument is true for the
troublesome VTL in the template which is easily messed up.